% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echamtools.R
\docType{package}
\name{echamtools-package}
\alias{echamtools}
\alias{echamtools-package}
\title{Generate summary warm/cold rain mask files from 3D ECHAM output}
\description{
Generate summary warm/cold rain mask files from 3D ECHAM output
}

