% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lsm_T63}
\alias{lsm_T63}
\title{T63 land/sea mask}
\format{A data frame with 6 variables:
\describe{
  \item{lon}{longitude; -180 < lon < 180}
  \item{lat}{latitude}
  \item{lsm}{land/sea mask}
}}
\source{
See the function generate_lsm() in R/lsm.R
}
\usage{
lsm_T63
}
\description{
A dataset containing the ECHAM land/sea mask at T63 resolution
}
\keyword{datasets}
