% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process.precip.profile.echam}
\alias{process.precip.profile.echam}
\title{Generate summary warm/cold rain mask files from 3D ECHAM output}
\usage{

  process.precip.profile.echam(datadir = "/work/bb0839/b380126/mpiesm-1.2.00p1/src/echam/experiments",
  experiment = "amip-rain-15", out.prefix = "", years = 1979:1983,
  ncores = 12, flux = TRUE, subsample = NULL)
}
\arguments{
\item{datadir}{Path to data; this directory should containt \code{experiment} as a subdirectory.}

\item{experiment}{Subdirectory of \code{datadir} that contains the experiment to process.}

\item{years}{Numeric.  Vector of years to process.}

\item{ncores}{Single number.  Number of children to use in processing the data.}
}
\value{
\code{NULL}
}
\description{
Prerequisite is ECHAM model output that contains rain and snow
vertical profiles as variables \code{aprlv_na} and \code{aprsv_na}.
Files should be created at monthly time frequency (but data time
frequency is arbitrary).  As a side effect of calling this
function, one RDS file per ECHAM file is created containing the 2D
precip mask at each time in the input file.  The precipitation mask
can take the values \code{dry}, \code{warm}, \code{cold}, and
\code{snow}.
}
\details{
At the end of processing, a summary file containing the number of
occurrences of dry/warm/cold/snow in each grid box is produced.
This file is named \code{experiment}.rds and is written in the
working directory.
}
